// js/stat-evaluation-player/src/generated/RotationPlayerStats.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { PlayDepthState } from "./PlayDepthState.ts";
import type { RoleState } from "./RoleState.ts";

export type RotationPlayerStats = { active_game_time: number, tracked_time: number, time_first_man: number, time_second_man: number, time_third_man: number, time_ambiguous_role: number, time_behind_play: number, time_level_with_play: number, time_ahead_of_play: number, longest_first_man_stint_time: number, first_man_stint_count: number, became_first_man_count: number, lost_first_man_count: number, current_role_state: RoleState, current_depth_state: PlayDepthState, };