// js/stat-evaluation-player/src/generated/RotationTeamEvent.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { RemoteIdTs } from "./RemoteIdTs.ts";

export type RotationTeamEvent = { time: number, frame: number, is_team_0: boolean, previous_first_man: RemoteIdTs, next_first_man: RemoteIdTs, };